cosmetic changes, cleaning old code

+++ src/main/java/org/pio/main/GameScreen.java
@@ -14,7 +14,7 @@ public class GameScreen extends JPanel {
 
     public static int screenWidth,screenHeight, intScreenWidth, intScreenHeight, intSidePanelStart;
 
-    public static final float SCALE=2;
+    public static final int SCALE=2;
     public static final int UNIT_SIZE =32;
     public static final float SCALED_UNIT_SIZE = UNIT_SIZE*SCALE;
